Shift XMLTransformer unit tests to new test module
